AAP-28425: Add tabs to Remotes, Access tab has a placeholder
